Make kebab case the default in @ConfigProperties
