+++ extensions/arc/deployment/src/main/java/io/quarkus/arc/deployment/ArcConfig.java
@@ -7,6 +7,7 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 
+import io.quarkus.arc.config.ConfigProperties;
 import io.quarkus.runtime.annotations.ConfigItem;
 import io.quarkus.runtime.annotations.ConfigRoot;
 
@@ -59,6 +60,13 @@ public class ArcConfig {
      */
     @ConfigItem(defaultValue = "true")
     public boolean removeFinalForProxyableMethods;
+
+    /**
+     * The default naming strategy for {@link ConfigProperties.NamingStrategy}. The allowed values are determined
+     * by that enum
+     */
+    @ConfigItem(defaultValue = "kebab-case")
+    public ConfigProperties.NamingStrategy configPropertiesDefaultNamingStrategy;
 
     public final boolean isRemoveUnusedBeansFieldValid() {
         return ALLOWED_REMOVE_UNUSED_BEANS_VALUES.contains(removeUnusedBeans.toLowerCase());
