+++ extensions/arc/deployment/src/test/java/io/quarkus/arc/test/configproperties/TypicalClassConfigPropertiesTest.java
@@ -27,7 +27,7 @@ public class TypicalClassConfigPropertiesTest {
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
                     .addClasses(DummyBean.class, DummyProperties.class)
                     .addAsResource(new StringAsset(
-                            "dummy.name=quarkus\ndummy.numbers=1,2,3,4\ndummy.boolWithDefault=true\ndummy.optionalInt=100"),
+                            "dummy.name=quarkus\ndummy.numbers=1,2,3,4\ndummy.bool-with-default=true\ndummy.optional-int=100"),
                             "application.properties"));
 
     @Inject
